requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astroid==3.3.9
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2024.2.2
click==8.1.8
cpplint==2.0.2
dill==0.3.9
ecdsa==0.19.1
fastapi==0.127.0
flasgger==0.9.7.1
Flask==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
mistune==3.1.3
packaging==24.2
passlib==1.7.4
platformdirs==4.3.7
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pylint==3.3.6
pytest==8.3.5
python-jose==3.5.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
starlette==0.50.0
tomlkit==0.13.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
Werkzeug==3.1.3
